backend/requirements.txt
google-generativeai
python-dotenv
Flask
requests
psycopg2-binary
Flask-Cors
weasyprint
Pillow
amadeus

